requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-core==2.23.4
alembic==1.13.2
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
pdfminer.six==20240706
PyJWT==2.9.0 